新增功能

+++ hosp-dao/src/main/java/cn/com/hosp/www/dao/mapper/TransportTaskMapper.java
@@ -26,7 +26,7 @@ public interface TransportTaskMapper extends BaseMapper<TransportTask> {
     List<TransportPatient> queryByPage(Map<String, Object> params);
 
 
-    int queryCount(Map<String, Object> params);
+    long queryCount(Map<String, Object> params);
 
     /**
      * 任务分派
